+++ pkg/providers/property/domain_validation_handler.go
@@ -183,11 +183,10 @@ func (v *validationHandler) buildInvalidateRequest() *domainownership.Invalidate
 	return &domainownership.InvalidateDomainsRequest{Domains: domainsToInvalidateSlice}
 }
 
-func (v *validationHandler) ensureValidationMethodsSupported() error {
+func (v *validationHandler) ensureCorrectValidationMethods() error {
 	var invalidDomains []string
 	for domainKey, domainDetails := range v.domainsToValidate {
-		validationMethod := domainDetails.validationMethod
-		if (validationMethod == string(domainownership.ValidationMethodSYSTEM) ||
+		if validationMethod := domainDetails.validationMethod; (validationMethod == string(domainownership.ValidationMethodSYSTEM) ||
 			validationMethod == string(domainownership.ValidationMethodMANUAL)) && domainDetails.validationStatus != "VALIDATED" {
 			invalidDomains = append(invalidDomains, domainKey.domainName)
 		}
@@ -195,6 +194,17 @@ func (v *validationHandler) ensureValidationMethodsSupported() error {
 
 	if len(invalidDomains) > 0 {
 		return fmt.Errorf("invalid validation method: the validation methods 'SYSTEM' and 'MANUAL' are not supported for validating new domains. Domains with invalid validation methods: %v", invalidDomains)
+	}
+
+	invalidDomains = []string{}
+	for planDomainKey, planDomainDetails := range v.planDomains {
+		if apiDomainDetails, ok := v.apiDomains[planDomainKey]; ok && apiDomainDetails.validationStatus == "VALIDATED" && planDomainDetails.validationMethod != apiDomainDetails.validationMethod {
+			invalidDomains = append(invalidDomains, planDomainKey.domainName)
+		}
+	}
+
+	if len(invalidDomains) > 0 {
+		return fmt.Errorf("validation method cannot be changed for existing domains. Domains with changed validation methods: %v", invalidDomains)
 	}
 
 	return nil